requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.9.5
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
async-lru==2.0.5
attrs==25.1.0
audioop-lts==0.2.1
azure-cognitiveservices-speech>=1.31.0
azure-common==1.1.28
azure-core==1.32.0
azure-cosmos==4.9.0
azure-identity==1.17.1
azure-keyvault-secrets==4.9.0
azure-monitor-opentelemetry-exporter==1.0.0b35
azure-search-documents==11.4.0
babel==2.17.0
beautifulsoup4==4.13.3
bidict==0.23.1
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
clipboard==0.0.4
colorama==0.4.6
comm==0.2.2
cryptography==44.0.2
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
eventlet==0.39.1
exa-py==1.8.5
executing==2.2.0
fastapi>=0.68.0
fastapi-cli==0.0.7
fastjsonschema==2.21.1
ffmpy==0.5.0
filelock==3.18.0
fixedint==0.1.6
Flask==3.1.0
flask-cors==5.0.1
Flask-SocketIO==5.5.1
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2025.3.0
gevent==24.11.1
gitdb==4.0.12
GitPython==3.1.44
gradio==5.23.0
gradio_client==1.8.0
greenlet==3.1.1
griffe==1.6.2
groovy==0.1.2
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.5
isodate==0.7.2
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.9.0
json5==0.10.0
jsonlines==4.0.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
linkify-it-py==2.0.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
mistune==3.1.2
msal==1.32.0
msal-extensions==1.3.1
msrest==0.7.1
multidict==6.1.0
narwhals==1.24.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.3.3
notebook_shim==0.2.4
numpy>=1.21.0
oauthlib==3.2.2
openai>=1.0.0
openai-agents==0.0.6
opentelemetry-api==1.31.0
opentelemetry-instrumentation==0.52b0
opentelemetry-instrumentation-asgi==0.52b0
opentelemetry-instrumentation-fastapi==0.52b0
opentelemetry-sdk==1.31.0
opentelemetry-semantic-conventions==0.52b0
opentelemetry-util-http==0.52b0
orjson==3.10.16
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.7
prometheus_client==0.21.1
prompt_toolkit==3.0.50
prompty==0.1.24
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.0
pycparser==2.22
pydantic>=2.0.0
pydantic_core==2.27.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.1
PyJWT==2.10.1
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv>=0.19.0
python-engineio==4.11.2
python-json-logger==3.3.0
python-multipart==0.0.20
python-socketio==5.12.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.3.0
referencing==0.36.2
requests>=2.26.0
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rich-toolkit==0.13.2
rpds-py==0.22.3
ruff==0.11.2
safehttpx==0.1.6
semantic-version==2.10.0
Send2Trash==1.8.3
setuptools==75.8.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.1
soupsieve==2.6
stack-data==0.6.3
starlette==0.46.1
streamlit==1.41.1
tabulate==0.9.0
tenacity==9.0.0
terminado==0.18.1
textual==2.1.2
tinycss2==1.4.0
toml==0.10.2
tomlkit==0.13.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.2
types-python-dateutil==2.9.0.20241206
types-requests==2.32.0.20250306
typing-extensions>=4.0.0
tzdata==2025.1
uc-micro-py==1.0.3
uri-template==1.3.0
urllib3==2.3.0
uvicorn>=0.15.0
uvloop==0.21.0
watchfiles==1.0.4
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets>=10.0
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.13
wrapt==1.17.2
wsproto==1.2.0
yarl==1.18.3
zipp==3.21.0
zope.event==5.0
zope.interface==7.2
# Speech-to-text services
boto3>=1.26.0
botocore>=1.29.0
/Users/apple/Desktop/Nirmitee/my-open-source/openai-voice-agents/requirements.txt